集成各服务子代理修复并重建：非 identical 615->560
- 修 dbmw DNFPacketTranslater 的 Event_Start 字段名（m_fieldA->m_eventCode）
- 修 monitor C5 工具链缺 glibc 头（dnf_monitor_cxx.sh 补 c5root include）
- 集成 statics/monitor/guild/dbmw 子代理的字段/控制流/栈布局修复
- monitor 223->185，guild 208->192，dbmw 3 个 DIFF->NEAR
- manager 仍 1（SendPacket 栈槽顺序 NEAR，见 attempts）

各服务非 identical：statics 40 / monitor 185 / guild 192 / dbmw 142 / manager 1

+++ dnf_decompile/source/DNFServer/GameServer/Monitor/DNFProhibitUser.cpp
@@ -64,7 +64,7 @@ void CDNFProhibitUser::SetUserConnectableTime(unsigned int dbid, short time, cha
 {
     if (!m_flag)
     {
-        m_remain = (unsigned short)time;
+        m_remain = time;
         m_dbid = dbid;
         m_channel = channel;
         m_flag = flag;
